refactor(nge-doc): group src into shared, core and ui zones
- src/shared: the pure, framework-free utils (frontmatter, slug, pages)
  shared by the runtime, the compiler and the SSR entry. The Node
  builder now ships this folder wholesale instead of maintaining a
  per-file whitelist in build-nge-builder.sh.
- src/core: types, manifest, assets, search, providers and services.
- src/ui: the routed component, the renderer and the layouts.

Import specifiers and the entry point follow; no behavior change.
