+++ projects/nge/doc/compiler/build.ts
@@ -1,8 +1,8 @@
 import { mkdirSync, writeFileSync } from 'node:fs'
 import { dirname } from 'node:path'
-import type { NgeDocMeta } from '../src/nge-doc'
-import type { NgeDocManifest } from '../src/manifest'
-import { parseFrontmatter } from '../src/frontmatter'
+import type { NgeDocMeta } from '../src/core/nge-doc'
+import type { NgeDocManifest } from '../src/core/manifest'
+import { parseFrontmatter } from '../src/shared/frontmatter'
 import { compileDocs } from './compile'
 import { DocFs, nodeFs } from './fs'
 import { DocGit, nodeGit } from './git'
