+++ projects/nge/doc/compiler/seo.spec.ts
@@ -1,4 +1,4 @@
-import type { NgeDocManifest } from '../src/manifest'
+import type { NgeDocManifest } from '../src/core/manifest'
 import { buildRobots, buildSitemap } from './seo'
 
 const manifest: NgeDocManifest = {
