+++ projects/nge/doc/index.ts
@@ -1,10 +1,10 @@
-export * from './src/nge-doc'
-export * from './src/assets'
-export * from './src/manifest'
-export * from './src/pages'
-export * from './src/search'
-export * from './src/nge-doc.module'
-export * from './src/nge-doc.providers'
-export * from './src/nge-doc.service'
-export * from './src/nge-doc-theme.service'
-export * from './src/renderer/renderer.component'
+export * from './src/core/nge-doc'
+export * from './src/core/assets'
+export * from './src/core/manifest'
+export * from './src/core/search'
+export * from './src/core/nge-doc.providers'
+export * from './src/core/nge-doc.service'
+export * from './src/core/nge-doc-theme.service'
+export * from './src/shared/pages'
+export * from './src/ui/nge-doc.module'
+export * from './src/ui/renderer/renderer.component'
