+++ projects/nge/doc/src/core/assets.spec.ts
