+++ projects/nge/doc/src/core/assets.ts
