+++ projects/nge/doc/src/core/manifest.spec.ts
