+++ projects/nge/doc/src/core/manifest.ts
