+++ projects/nge/doc/src/core/nge-doc.ts
