+++ projects/nge/doc/src/core/search.ts
