+++ projects/nge/doc/src/shared/pages.ts
@@ -1,11 +1,11 @@
-import type { NgeDocLink } from './nge-doc'
+import type { NgeDocLink } from '../core/nge-doc'
 
 /**
  * Content pages (an `href` plus a renderer) of a navigation tree, depth-first.
  * Separators, pure grouping links and external links (no renderer) are
- * excluded. Framework-free and shared by the runtime, the compiler and the
- * SSR entry (kept at builder runtime by build-nge-builder.sh, like
- * {@link ./frontmatter}).
+ * excluded. Shared by the runtime, the compiler and the SSR entry; `src/shared`
+ * holds the pure, framework-free utils and is the only src folder the Node
+ * builder ships (the link type import is type-only, erased at compile time).
  */
 export function contentPages(pages: NgeDocLink[]): NgeDocLink[] {
   const out: NgeDocLink[] = []
