+++ projects/nge/doc/src/ui/layouts/default/default-layout.component.ts
@@ -12,8 +12,8 @@ import {
 import { DatePipe } from '@angular/common'
 import { Router } from '@angular/router'
 import { NgeDocRendererComponent } from '../../renderer/renderer.component'
-import { NgeDocLink } from '../../nge-doc'
-import { NgeDocService } from '../../nge-doc.service'
+import { NgeDocLink } from '../../../core/nge-doc'
+import { NgeDocService } from '../../../core/nge-doc.service'
 import { BreadcrumbComponent } from './breadcrumb/breadcrumb.component'
 import { HeaderComponent } from './header/header.component'
 import { PageActionsComponent } from './page-actions/page-actions.component'
