+++ projects/nge/doc/src/ui/layouts/default/pager/pager.component.scss
