+++ projects/nge/doc/src/ui/layouts/default/pager/pager.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core'
 import { RouterLink } from '@angular/router'
-import { NgeDocService } from '../../../nge-doc.service'
+import { NgeDocService } from '../../../../core/nge-doc.service'
 
 @Component({
   selector: 'nge-doc-pager',
