+++ projects/nge/doc/src/ui/layouts/default/search/search-palette.component.html
