+++ projects/nge/doc/src/ui/layouts/default/sidenav/sidenav.component.scss
