+++ projects/nge/doc/src/ui/layouts/default/toc/toc.component.scss
