+++ projects/nge/doc/src/ui/nge-doc.component.scss
